backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.1.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
pydantic==2.4.2
pydantic-settings==2.0.3
onnxruntime-gpu==1.16.0
insightface==0.7.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
ultralytics==8.0.200
torch==2.0.1
torchvision==0.15.2
scikit-image==0.21.0
scikit-learn==1.3.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
httpx==0.25.0
aiofiles==23.2.1
websockets==11.0.3
redis==5.0.0
celery==5.3.1
supervision==0.17.1